requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1
twilio==7.0.0
gunicorn==20.1.0 